Upload Jupyter Notebook: Farmburg
Codecademy,

Thank you for your time and consideration. This was by far the most enjoyable application/exercise I've yet had to complete for a potential employer. I sincerely hope to continue to be considered for a coach/mentor position with Codecademy. I know first hand the value your service provides to individuals looking to learn how to program. 

Christian Golden